requirements.txt
asgiref==3.8.1
blinker==1.4
certifi==2024.2.2
charset-normalizer==3.3.2
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.3
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-filter==23.5
django-js-asset==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.13.0
greenlet==3.0.3
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
inflection==0.5.1
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.3
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
more-itertools==8.10.0
msgpack==1.0.7
netifaces==0.11.0
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==24.0
pillow==10.2.0
platformdirs==3.11.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
pycairo==1.20.1
PyGObject==3.42.1
PyJWT==2.8.0
pynvim==0.5.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
SecretStorage==3.3.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
systemd-python==234
typing_extensions==4.10.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.24.6
wadllib==1.3.6
wcwidth==0.2.10
zipp==1.0.0
